# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /checkout:
    post:
      operationId: checkout
      tags: [orders]
      summary: Place an order from the current cart
      security: [{ clerkSession: [] }]
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckoutRequest"
      responses:
        "201":
          description: Order placed successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerOrder"
        "400":
          description: Cart is empty or a product is unavailable
        "401":
          description: Authentication required
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List the signed-in user's orders
      security: [{ clerkSession: [] }]
      responses:
        "200":
          description: Customer orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CustomerOrder"
        "401":
          description: Authentication required
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get a single order with its line items
      security: [{ clerkSession: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Order detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerOrderDetail"
        "401":
          description: Authentication required
        "404":
          description: Order not found
  /orders/{id}/cancel:
    post:
      operationId: cancelOrder
      tags: [orders]
      summary: Cancel an order (only allowed before it is shipped)
      security: [{ clerkSession: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Order cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerOrder"
        "400":
          description: Order cannot be cancelled in its current state
        "401":
          description: Authentication required
        "404":
          description: Order not found
  /merchant/summary:
    get:
      operationId: getMerchantSummary
      tags: [merchant]
      summary: Get the signed-in merchant workspace summary
      security: [{ clerkSession: [] }]
      responses:
        "200":
          description: Merchant summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MerchantSummary"
        "401":
          description: Authentication required
  /merchant/store:
    get:
      operationId: getMerchantStore
      tags: [merchant]
      summary: Get the signed-in merchant store
      security: [{ clerkSession: [] }]
      responses:
        "200":
          description: Merchant store
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MerchantStore"
        "401":
          description: Authentication required
        "404":
          description: Store not found
    post:
      operationId: createMerchantStore
      tags: [merchant]
      summary: Create the signed-in merchant store
      security: [{ clerkSession: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMerchantStoreRequest"
      responses:
        "201":
          description: Merchant store created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MerchantStore"
        "400":
          description: Invalid store
        "401":
          description: Authentication required
        "409":
          description: Store already exists
  /merchant/products:
    get:
      operationId: listMerchantProducts
      tags: [merchant]
      summary: List products owned by the signed-in merchant
      security: [{ clerkSession: [] }]
      responses:
        "200":
          description: Merchant products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MerchantProduct"
        "401":
          description: Authentication required
    post:
      operationId: createMerchantProduct
      tags: [merchant]
      summary: Create a product in the signed-in merchant store
      security: [{ clerkSession: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMerchantProductRequest"
      responses:
        "201":
          description: Product created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MerchantProduct"
        "400":
          description: Invalid product
        "401":
          description: Authentication required
        "404":
          description: Store or category not found
  /merchant/products/{id}:
    patch:
      operationId: updateMerchantProduct
      tags: [merchant]
      summary: Update a product owned by the signed-in merchant
      security: [{ clerkSession: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMerchantProductRequest"
      responses:
        "200":
          description: Product updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MerchantProduct"
        "400":
          description: Invalid product
        "401":
          description: Authentication required
        "404":
          description: Product not found
    delete:
      operationId: deleteMerchantProduct
      tags: [merchant]
      summary: Archive a product owned by the signed-in merchant
      security: [{ clerkSession: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Product archived
        "401":
          description: Authentication required
        "404":
          description: Product not found
...
  /merchant/orders:
    get:
      operationId: listMerchantOrders
      tags: [merchant]
      summary: List orders for the signed-in merchant store
      security: [{ clerkSession: [] }]
      responses:
        "200":
          description: Merchant orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MerchantOrder"
        "401":
          description: Authentication required
  /merchant/orders/{id}/status:
    patch:
      operationId: updateMerchantOrderStatus
      tags: [merchant]
      summary: Advance the fulfillment status of a merchant order (pending → processing → shipped → delivered)
      security: [{ clerkSession: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMerchantOrderStatusRequest"
      responses:
        "200":
          description: Order status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MerchantOrder"
        "400":
          description: Invalid status transition
        "401":
          description: Authentication required
        "404":
          description: Order not found
...
components:
  securitySchemes:
    clerkSession:
      type: apiKey
      in: cookie
      name: __session
  schemas:
...
    CheckoutRequest:
      type: object
      properties:
        couponCode:
          type: string
          description: Optional coupon code to apply at checkout
    CustomerOrder:
      type: object
      properties:
        id:
          type: string
        status:
          type: string
        total:
          type: number
        cashbackAmount:
          type: number
        discountAmount:
          type: number
        couponCode:
          type: ["string", "null"]
        itemsCount:
          type: number
        deliveredAt:
          type: ["string", "null"]
        completedAt:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, status, total, cashbackAmount, discountAmount, couponCode, itemsCount, deliveredAt, completedAt, createdAt, updatedAt]
    CustomerOrderItem:
      type: object
      properties:
        id:
          type: string
        productId:
          type: string
        storeId:
          type: string
        name:
          type: string
        imageUrl:
          type: string
        price:
          type: number
        originalPrice:
          type: number
        cashbackPercent:
          type: number
        cashbackAmount:
          type: number
        quantity:
          type: number
      required: [id, productId, storeId, name, imageUrl, price, originalPrice, cashbackPercent, cashbackAmount, quantity]
    CustomerOrderDetail:
      type: object
      properties:
        id:
          type: string
        status:
          type: string
        total:
          type: number
        cashbackAmount:
          type: number
        discountAmount:
          type: number
        couponCode:
          type: ["string", "null"]
        itemsCount:
          type: number
        deliveredAt:
          type: ["string", "null"]
        completedAt:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/CustomerOrderItem"
      required: [id, status, total, cashbackAmount, discountAmount, couponCode, itemsCount, deliveredAt, completedAt, createdAt, updatedAt, items]
...
    MerchantStore:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        slug: { type: string }
        description: { type: string }
        logoUrl: { type: string }
        status: { type: string }
        createdAt: { type: string }
        updatedAt: { type: string }
      required: [id, name, slug, description, logoUrl, status, createdAt, updatedAt]
    CreateMerchantStoreRequest:
      type: object
      properties:
        name: { type: string }
        description: { type: string }
        logoUrl: { type: string }
      required: [name]
    MerchantProduct:
      type: object
      properties:
        id: { type: string }
        storeId: { type: string }
        categoryId: { type: string }
        categoryName: { type: string }
        name: { type: string }
        description: { type: string }
        brand: { type: string }
        price: { type: number }
        originalPrice: { type: number }
        cashbackPercent: { type: number }
        imageUrl: { type: string }
        stock: { type: number }
        available: { type: boolean }
        status: { type: string }
        createdAt: { type: string }
        updatedAt: { type: string }
      required: [id, storeId, categoryId, categoryName, name, description, brand, price, originalPrice, cashbackPercent, imageUrl, stock, available, status, createdAt, updatedAt]
    CreateMerchantProductRequest:
      type: object
      properties:
        categoryId: { type: string }
        name: { type: string }
        description: { type: string }
        brand: { type: string }
        price: { type: number }
        originalPrice: { type: number }
        cashbackPercent: { type: number }
        imageUrl: { type: string }
        stock: { type: number }
        available: { type: boolean }
      required: [categoryId, name, brand, price, originalPrice, cashbackPercent, imageUrl, stock, available]
    UpdateMerchantProductRequest:
      type: object
      properties:
        categoryId: { type: string }
        name: { type: string }
        description: { type: string }
        brand: { type: string }
        price: { type: number }
        originalPrice: { type: number }
        cashbackPercent: { type: number }
        imageUrl: { type: string }
        stock: { type: number }
        available: { type: boolean }
        status: { type: string }
    MerchantOrder:
      type: object
      properties:
        id: { type: string }
        storeId: { type: string }
        customerId: { type: string }
        total: { type: number }
        cashback: { type: number }
        itemsCount: { type: number }
        status: { type: string }
        createdAt: { type: string }
      required: [id, storeId, customerId, total, cashback, itemsCount, status, createdAt]
...
    UpdateMerchantOrderStatusRequest:
      type: object
      properties:
        status:
          type: string
          enum: [processing, shipped, delivered]
      required: [status]
    MerchantSummary:
      type: object
      properties:
        store:
          anyOf:
            - $ref: "#/components/schemas/MerchantStore"
            - type: "null"
        productCount: { type: number }
        activeProductCount: { type: number }
        orderCount: { type: number }
        grossSales: { type: number }
        cashbackIssued: { type: number }
        pendingOrders: { type: number }
      required: [store, productCount, activeProductCount, orderCount, grossSales, cashbackIssued, pendingOrders]
...