# lib/api-zod/src/generated/types/customerOrderDetail.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CustomerOrderItem } from './customerOrderItem';

export interface CustomerOrderDetail {
  id: string;
  status: string;
  total: number;
  cashbackAmount: number;
  discountAmount: number;
  /** @nullable */
  couponCode: string | null;
  itemsCount: number;
  /** @nullable */
  deliveredAt: string | null;
  /** @nullable */
  completedAt: string | null;
  createdAt: string;
  updatedAt: string;
  items: CustomerOrderItem[];
}