# lib/api-zod/src/generated/types/updateMerchantOrderStatusRequest.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateMerchantOrderStatusRequestStatus } from './updateMerchantOrderStatusRequestStatus';

export interface UpdateMerchantOrderStatusRequest {
  status: UpdateMerchantOrderStatusRequestStatus;
}